@@ -94,7 +94,7 @@ func (p *Processors) ProcessGroupMsgReject(data *dto.GroupMsgRejectEvent) error 
 		noticeMap := structToMap(notice)
 
 		//上报信息到onebotv11应用端(正反ws)
-		p.BroadcastMessageToAll(noticeMap)
+		go p.BroadcastMessageToAll(noticeMap, p.Apiv2, data)
 
 		// 转换appid
 		AppIDString := strconv.FormatUint(p.Settings.AppID, 10)
@@ -206,7 +206,7 @@ func (p *Processors) ProcessGroupMsgReject(data *dto.GroupMsgRejectEvent) error 
 			// Convert OnebotGroupMessage to map and send
 			groupMsgMap := structToMap(groupMsg)
 			//上报信息到onebotv11应用端(正反ws)
-			p.BroadcastMessageToAll(groupMsgMap)
+			go p.BroadcastMessageToAll(groupMsgMap, p.Apiv2, data)
 
 			// 转换appid
 			AppIDString := strconv.FormatUint(p.Settings.AppID, 10)
