@@ -3,6 +3,7 @@ package Processor
 
 import (
 	"fmt"
+	"log"
 	"strconv"
 	"time"
 
@@ -131,7 +132,7 @@ func (p *Processors) ProcessGuildNormalMessage(data *dto.WSMessageData) error {
 		msgMap := structToMap(onebotMsg)
 
 		//上报信息到onebotv11应用端(正反ws)
-		p.BroadcastMessageToAll(msgMap)
+		go p.BroadcastMessageToAll(msgMap, p.Apiv2, data)
 	} else {
 		// GlobalChannelToGroup为true时的处理逻辑
 		//将频道转化为一个群
@@ -191,10 +192,17 @@ func (p *Processors) ProcessGuildNormalMessage(data *dto.WSMessageData) error {
 		// 构造echostr，包括AppID，原始的s变量和当前时间戳
 		echostr := fmt.Sprintf("%s_%d_%d", AppIDString, s, currentTimeMillis)
 		//映射str的messageID到int
-		messageID64, err := idmap.StoreIDv2(data.ID)
-		if err != nil {
-			mylog.Printf("Error storing ID: %v", err)
-			return nil
+		var messageID64 int64
+		if config.GetMemoryMsgid() {
+			messageID64, err = echo.StoreCacheInMemory(data.ID)
+			if err != nil {
+				log.Fatalf("Error storing ID: %v", err)
+			}
+		} else {
+			messageID64, err = idmap.StoreCachev2(data.ID)
+			if err != nil {
+				log.Fatalf("Error storing ID: %v", err)
+			}
 		}
 		messageID := int(messageID64)
 		// 如果在Array模式下, 则处理Message为Segment格式
@@ -294,6 +302,8 @@ func (p *Processors) ProcessGuildNormalMessage(data *dto.WSMessageData) error {
 		echo.AddMsgType(AppIDString, ChannelID64, "guild")
 		//懒message_id池
 		echo.AddLazyMessageId(strconv.FormatInt(ChannelID64, 10), data.ID, time.Now())
+		//测试
+		echo.AddLazyMessageId(data.ChannelID, data.ID, time.Now())
 		//懒message_id池
 		//echo.AddLazyMessageId(strconv.FormatInt(userid64, 10), data.ID, time.Now())
 		//echo.AddLazyMessageIdv2(strconv.FormatInt(ChannelID64, 10), strconv.FormatInt(userid64, 10), data.ID, time.Now())
@@ -305,7 +315,7 @@ func (p *Processors) ProcessGuildNormalMessage(data *dto.WSMessageData) error {
 		groupMsgMap := structToMap(groupMsg)
 
 		//上报信息到onebotv11应用端(正反ws)
-		p.BroadcastMessageToAll(groupMsgMap)
+		go p.BroadcastMessageToAll(groupMsgMap, p.Apiv2, data)
 	}
 
 	return nil
