@@ -107,7 +107,7 @@ func (p *Processors) ProcessInlineSearch(data *dto.WSInteractionData) error {
 		noticeMap := structToMap(notice)
 
 		//上报信息到onebotv11应用端(正反ws)
-		p.BroadcastMessageToAll(noticeMap)
+		go p.BroadcastMessageToAll(noticeMap, p.Apiv2, data)
 
 		// 转换appid
 		AppIDString := strconv.FormatUint(p.Settings.AppID, 10)
@@ -221,7 +221,7 @@ func (p *Processors) ProcessInlineSearch(data *dto.WSInteractionData) error {
 			// Convert OnebotGroupMessage to map and send
 			groupMsgMap := structToMap(groupMsg)
 			//上报信息到onebotv11应用端(正反ws)
-			p.BroadcastMessageToAll(groupMsgMap)
+			go p.BroadcastMessageToAll(groupMsgMap, p.Apiv2, data)
 
 			// 转换appid
 			AppIDString := strconv.FormatUint(p.Settings.AppID, 10)
@@ -300,7 +300,7 @@ func (p *Processors) ProcessInlineSearch(data *dto.WSInteractionData) error {
 			// Convert OnebotGroupMessage to map and send
 			privateMsgMap := structToMap(privateMsg)
 			//上报信息到onebotv11应用端(正反ws)
-			p.BroadcastMessageToAll(privateMsgMap)
+			go p.BroadcastMessageToAll(privateMsgMap, p.Apiv2, data)
 
 			// 转换appid
 			AppIDString := strconv.FormatUint(p.Settings.AppID, 10)
@@ -361,7 +361,7 @@ func (p *Processors) ProcessInlineSearch(data *dto.WSInteractionData) error {
 			msgMap := structToMap(onebotMsg)
 
 			//上报信息到onebotv11应用端(正反ws)
-			p.BroadcastMessageToAll(msgMap)
+			go p.BroadcastMessageToAll(msgMap, p.Apiv2, data)
 
 			// TODO: 实现eventid
 		}
